requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
Eel==0.16.0
et-xmlfile==1.1.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
JPype1==1.5.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tabula-py==2.9.3
tzdata==2024.1
urllib3==2.2.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
